requirements.txt
# Production-ready requirements for BRICKIT

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Supabase
supabase>=2.3.0
postgrest>=0.16.0
gotrue>=2.4.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.0

# System Monitoring
psutil>=5.9.0

# Cache (optional)
redis>=5.0.0
hiredis>=2.2.0

# Logging
structlog>=23.2.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Production Server
gunicorn>=21.2.0

# Environment Management
python-dotenv>=1.0.0

# Data Validation
email-validator>=2.1.0

# Rate Limiting
slowapi>=0.1.9

# API Documentation
fastapi-users>=12.1.0

# Background Tasks
celery>=5.3.0

# File Upload
aiofiles>=23.2.0

# JSON Schema
jsonschema>=4.20.0

# Time Zone
pytz>=2023.3

# Development Database (SQLite alternative)
aiosqlite>=0.19.0
